+++ storage/news.go
@@ -15,10 +15,10 @@ type NewsStorage interface {
 	Close() error
 
 	/** Posts methods **/
-	GetMetaOfPosts(interface{}, int, int, string, []string) ([]models.Post, int, error)
-	GetFullPosts(interface{}, int, int, string, []string) ([]models.Post, int, error)
-	GetMetaOfTopics(interface{}, int, int, string, []string) ([]models.Topic, int, error)
-	GetFullTopics(interface{}, int, int, string, []string) ([]models.Topic, int, error)
+	GetMetaOfPosts(models.MongoQuery, int, int, string, []string) ([]models.Post, int, error)
+	GetFullPosts(models.MongoQuery, int, int, string, []string) ([]models.Post, int, error)
+	GetMetaOfTopics(models.MongoQuery, int, int, string, []string) ([]models.Topic, int, error)
+	GetFullTopics(models.MongoQuery, int, int, string, []string) ([]models.Topic, int, error)
 }
 
 // NewMongoStorage initializes the storage connected to Mongo database
@@ -38,20 +38,7 @@ func (m *MongoStorage) Close() error {
 }
 
 // GetDocuments ...
-func (m *MongoStorage) GetDocuments(qs interface{}, limit int, offset int, sort string, collection string, documents interface{}) (count int, err error) {
-	var q models.MongoQuery
-
-	_qs, ok := qs.(string)
-
-	if ok {
-		err = models.GetQuery(_qs, &q)
-
-		if err != nil {
-			return 0, m.NewStorageError(err, "GetDocuments", "storage.mongo_storage.get_documents.parse_query_error")
-		}
-
-		qs = q
-	}
+func (m *MongoStorage) GetDocuments(qs models.MongoQuery, limit int, offset int, sort string, collection string, documents interface{}) (count int, err error) {
 
 	err = m.db.DB(utils.Cfg.MongoDBSettings.DBName).C(collection).Find(qs).Limit(limit).Skip(offset).Sort(sort).All(documents)