+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### 1.0.2
-- Add AmazonSES service to send mails
+- Return leading_image_portrait field in post model.
+- Add AmazonSES service to send mails.
+- Integrate with circleci.
+- Add Cache-Control in response header and simply cors setting
+- Add CORS in response header for production environment
+- Only allow to query `published` posts and topics in production environment
 
 ### 1.0.1
 - Fix typo. Agolia to Algolia.